fix(work-design): close source qualification projections
