+++ package.json
@@ -123,9 +123,9 @@
     "test:work-design-policy-replay": "node scripts/require-shifu.mjs test:work-design-policy-replay && node --test scripts/check-work-design-policy-replay.test.mjs",
     "check:work-design-open-card": "node scripts/require-shifu.mjs check:work-design-open-card && node framework/work-design-open-card/tooling/check-work-design-open-card.mjs",
     "test:work-design-open-card": "node scripts/require-shifu.mjs test:work-design-open-card && node --test scripts/check-work-design-open-card.test.mjs",
-    "fix:work-design-outcome-history": "node scripts/require-shifu.mjs fix:work-design-outcome-history && uv run --project framework/core --frozen ruff format framework/core/src/python/kungfu/assignment_orchestration.py framework/core/src/python/kungfu/workspace_federation.py framework/core/tests/python/test_assignment_orchestration.py framework/core/tests/python/test_workspace_federation.py",
+    "fix:work-design-outcome-history": "node scripts/require-shifu.mjs fix:work-design-outcome-history && uv run --project framework/core --frozen ruff format framework/core/src/python/kungfu/assignment_orchestration.py framework/core/src/python/kungfu/workspace_federation.py framework/core/tests/python/test_assignment_orchestration.py framework/core/tests/python/test_workspace_federation.py framework/core/tests/python/test_work_design_outcome_history.py",
     "fix:work-design-outcome-js": "node scripts/require-shifu.mjs fix:work-design-outcome-js && biome check --write --unsafe framework/work-design-advisor/src/work-design-advisor.mjs framework/work-design-advisor/tooling/work-design-advisor-contract.mjs framework/work-design-open-card/src/work-design-open-card.mjs framework/work-design-policy-replay/src/index.mjs scripts/check-work-design-advisor.test.mjs scripts/check-work-design-open-card.test.mjs scripts/check-work-design-policy-replay.test.mjs",
-    "test:work-design-outcome-history": "node scripts/require-shifu.mjs test:work-design-outcome-history && PYTHONPATH=framework/core/src/python:${KUNGFU_NATIVE_PATH:-framework/core/build/Release} uv run --project framework/core --frozen ruff format --check framework/core/src/python/kungfu/assignment_orchestration.py framework/core/src/python/kungfu/workspace_federation.py framework/core/tests/python/test_assignment_orchestration.py framework/core/tests/python/test_workspace_federation.py && PYTHONPATH=framework/core/src/python:${KUNGFU_NATIVE_PATH:-framework/core/build/Release} uv run --project framework/core --frozen pytest -q framework/core/tests/python/test_assignment_orchestration.py framework/core/tests/python/test_workspace_federation.py && node --test scripts/check-work-design-advisor.test.mjs scripts/check-work-design-open-card.test.mjs scripts/check-work-design-policy-replay.test.mjs",
+    "test:work-design-outcome-history": "node scripts/require-shifu.mjs test:work-design-outcome-history && PYTHONPATH=framework/core/src/python:${KUNGFU_NATIVE_PATH:-framework/core/build/Release} uv run --project framework/core --frozen ruff format --check framework/core/src/python/kungfu/assignment_orchestration.py framework/core/src/python/kungfu/workspace_federation.py framework/core/tests/python/test_assignment_orchestration.py framework/core/tests/python/test_workspace_federation.py framework/core/tests/python/test_work_design_outcome_history.py && PYTHONPATH=framework/core/src/python:${KUNGFU_NATIVE_PATH:-framework/core/build/Release} uv run --project framework/core --frozen pytest -q framework/core/tests/python/test_assignment_orchestration.py framework/core/tests/python/test_workspace_federation.py framework/core/tests/python/test_work_design_outcome_history.py && node --test scripts/check-work-design-advisor.test.mjs scripts/check-work-design-open-card.test.mjs scripts/check-work-design-policy-replay.test.mjs",
     "work-design:open-card-preflight": "node scripts/require-shifu.mjs work-design:open-card-preflight && node framework/work-design-open-card/tooling/open-card-preflight.mjs",
     "work-design:feedback": "node scripts/require-shifu.mjs work-design:feedback && node framework/work-design-policy-replay/tooling/check-work-design-policy-replay.mjs feedback",
     "check:project-cut-composition": "node scripts/require-shifu.mjs check:project-cut-composition && node scripts/check-project-cut-composition.mjs",
